added log command

+++ Commands/Debuging.js
@@ -15,17 +15,37 @@ module.exports = {
 
             message.channel.send(Error);
         }
-        else if(args[0] === 'kill')
+        else if(args[0] === 'kill' && message.member.roles.cache.has('748515939821158443'))
         {
             message.reply("Resetting...");
             client.destroy();
             client.login(process.env.token);
         }
+        else if(args[0] === 'Log' && message.member.roles.cache.has('748515939821158443'))
+        {
+            try
+            {
+                const NoErr = new Discord.MessageEmbed();
+                NoErr.setTitle('<:greentick:750751680613843105> No Errors !');
+                NoErr.setColor(0x32ba4b);
+
+                message.channel.send(NoErr);
+            }
+            catch(error)
+            {
+                const Err = new Discord.MessageEmbed();
+                Err.setTitle('<:redtick:750751681175748608> **Error Found !**');
+                Err.setDescription(error);
+                error.setColor(0xf53d3d);
+
+                message.channel.send(Err);
+            }
+        }
         else if(message.member.roles.cache.has('748515939821158443') && !args.lenght)
         {
             const DebugCommandsList = new Discord.MessageEmbed();
             DebugCommandsList.setTitle('**List of Debugging commands: **');
-            DebugCommandsList.addField('Commands: ', 'N/A', false);
+            DebugCommandsList.addField('Commands: ', '`Kill`, `Log`', false);
             DebugCommandsList.setColor(0x2abbf5);
 
             message.channel.send(DebugCommandsList);
